Module AG

+++ src/main/java/com/example/solimus/dtos/syndic/meeting/AGCardDTO.java
@@ -5,7 +5,7 @@ import java.time.LocalTime;
 import lombok.*;
 
 // ===== DTO CARTE ASSEMBLEE GENERALE (liste/dashboard) =====
-// Represente une carte AG affichee dans la liste, avec toutes les infos resumees
+// Représente une carte AG affichée dans la liste, avec toutes les infos resumées
 @Getter
 @Setter
 @Builder
@@ -15,12 +15,12 @@ public class AGCardDTO {
 
     private Long id;                    // identifiant technique de la reunion
 
-    private String title;               // titre de l'AG saisi a la creation
-    private String residenceName;       // nom de la residence concernee
+    private String title;               // titre de l'AG saisi à la creation
+    private String residenceName;       // nom de la residence concernée
     private Long residenceId;           // id residence, utile pour navigation cote front
 
     private String status;              // valeur technique du MeetingStatus (ex: "UPCOMING")
-    private String statusLabel;         // libelle affichable du statut (ex: "A venir")
+    private String statusLabel;         // libelle affichable du statut (ex: "à venir")
 
     private String type;                // valeur technique du MeetingType (ex: "ORDINARY")
     private String typeLabel;           // libelle affichable du type (ex: "Ordinaire")
@@ -28,12 +28,12 @@ public class AGCardDTO {
     private LocalDate meetingDate;      // date de la reunion
     private LocalTime startTime;        // heure de debut
 
-    private String location;            // lieu de la reunion
+    private String location;            // lieu de la réunion
 
-    private int presentCount;           // X : nombre de participants ayant signe (hasSigned = true)
-    private int totalParticipants;      // Y : nombre total de participants convoques a l'AG
-    private double participationRate;   // % de tantieme present, calcule via tantiemeSnapshot des presents
+    private long presentCount;           // X : nombre de participants ayant signé (hasSigned = true)
+    private long totalParticipants;      // Y : nombre total de participants convoqués à l'AG
+    private double participationRate;   // % de tantième présent, calculé via tantièmeSnapshot des présents
 
-    private int resolutionsCount;       // nombre de points a l'ordre du jour (agendaItems.size())
-    private int documentsCount;         // nombre de documents lies a l'AG
+    private long resolutionsCount;       // nombre de points à l'ordre du jour (agendaItems.size())
+    private long documentsCount;         // nombre de documents liés à l'AG
 }
